fix(cloud): harden Azure adoption recovery
